requirements.txt
django
djangorestframework
psycopg2
Pillow
cloudinary
dj3-cloudinary-storage